+++ craft_cli/utils.py
@@ -1,0 +1,22 @@
+# Copyright 2024 Canonical Ltd.
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU Lesser General Public
+# License version 3 as published by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with this program; if not, write to the Free Software Foundation,
+# Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+
+"""Utility functions for craft_cli."""
+
+
+def humanize_list(values: list[str], conjunction: str = "and") -> str:
+    """Convert a collection of values into a string that lists the values."""
+    start = ", ".join(values[:-1])
+    return f"{start}, {conjunction} {values[-1]}"