+++ tests/unit/test_messages_emitter.py
@@ -20,9 +20,8 @@ import logging
 import sys
 from unittest import mock
 from unittest.mock import call, patch
-from typing import cast, Callable
+from typing import Any, cast, Callable
 
-import craft_cli
 import pytest
 import pytest_mock
 
@@ -87,6 +86,34 @@ def get_initiated_emitter(tmp_path, monkeypatch):
             return emitter
 
         yield func
+
+def emitter_methods(init: bool, stop: bool = True, exclude: list[str] = []) -> list[Callable[..., Any]]:
+    """Provide a list of all public methods on an Emitter object.
+
+    :param init: Whether or not to initialize the emitter first
+    :param stop: Whether or not to stop the emitter after initialization. Does nothing if init is False.
+        Defaults to true.
+    :param exclude: A list of method names to exclude from the final output. Defaults to empty.
+    """
+    emitter = Emitter()
+    if init:
+        emitter.init(EmitterMode.QUIET, "testappname", "default greeting")
+        if stop:
+            emitter.ended_ok()
+
+    # Collect all the public methods in Emitter
+    all_methods = [item for item in dir(Emitter) if item[0] != "_"]
+
+    # Filter out from the exclusion list
+    all_methods = [item for item in all_methods if item not in exclude]
+
+    # Get the actual attributes
+    all_methods = [getattr(emitter, item) for item in all_methods]
+
+    # Filter out anything that isn't actually a method
+    all_methods = [item for item in all_methods if isinstance(item, Callable)]
+
+    return all_methods
 
 
 # -- tests for init and setting/getting mode
@@ -180,12 +207,9 @@ def test_init_developer_modes(mode, tmp_path, monkeypatch):
     (handler,) = [x for x in logger.handlers if isinstance(x, _Handler)]
     assert handler.mode == mode
 
-
-@pytest.mark.parametrize("method_name", [x for x in dir(Emitter) if x[0] != "_" and x != "init"])
-def test_needs_init(method_name):
+@pytest.mark.parametrize("method", emitter_methods(init=False, exclude=["init"]))
+def test_needs_init(method):
     """Check that calling other methods needs emitter first to be initiated."""
-    emitter = Emitter()
-    method = getattr(emitter, method_name)
     with pytest.raises(RuntimeError, match="Emitter needs to be initiated first"):
         method()
 
@@ -775,17 +799,9 @@ def test_ended_double_after_error(get_initiated_emitter):
     emitter.ended_ok()
     assert emitter.printer_calls == []
 
-
-@pytest.mark.parametrize(
-    "method_name",
-    [x for x in dir(Emitter) if x[0] != "_" and x not in ("init", "ended_ok", "error")],
-)
-def test_needs_being_active(get_initiated_emitter, method_name):
+@pytest.mark.parametrize("method", emitter_methods(init=True, exclude=["init", "ended_ok", "error"]))
+def test_needs_being_active(method):
     """Check that calling public methods needs emitter to not be stopped."""
-    emitter = get_initiated_emitter(EmitterMode.QUIET)
-    emitter.ended_ok()
-
-    method = getattr(emitter, method_name)
     with pytest.raises(RuntimeError, match="Emitter is stopped already"):
         method()